+++ invesdwin-util-parent/invesdwin-util/src/main/java/de/invesdwin/util/math/expression/lambda/IEvaluateIntegerKey.java
@@ -6,6 +6,6 @@ public interface IEvaluateIntegerKey {
     /**
      * Double.NaN is interpreted as 0.
      */
-    int evaluateInteger(int key);
+    int evaluateInteger(long key);
 
 }
