+++ invesdwin-util-parent/invesdwin-util/src/main/java/de/invesdwin/util/math/expression/lambda/IEvaluateNegativeReasonKey.java
@@ -6,6 +6,6 @@ public interface IEvaluateNegativeReasonKey {
     /**
      * Double.NaN is interpreted as null.
      */
-    Boolean evaluateBooleanNullable(int key);
+    Boolean evaluateBooleanNullable(long key);
 
 }
