+++ invesdwin-util-parent/invesdwin-util/src/main/java/de/invesdwin/util/math/expression/lambda/throwing/ThrowingEvaluateBooleanKey.java
@@ -16,7 +16,7 @@ public class ThrowingEvaluateBooleanKey implements IEvaluateBooleanKey {
     }
 
     @Override
-    public boolean evaluateBoolean(final int key) {
+    public boolean evaluateBoolean(final long key) {
         throw Throwables.propagate(throwable);
     }
