+++ invesdwin-util-parent/invesdwin-util/src/main/java/de/invesdwin/util/math/expression/lambda/throwing/ThrowingEvaluateIntegerKey.java
@@ -16,7 +16,7 @@ public class ThrowingEvaluateIntegerKey implements IEvaluateIntegerKey {
     }
 
     @Override
-    public int evaluateInteger(final int key) {
+    public int evaluateInteger(final long key) {
         throw Throwables.propagate(throwable);
     }
